+++ public/sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'wos-bdc-pwa-v1.95.3';
+const CACHE_NAME = 'wos-bdc-pwa-v1.95.4';
 const ASSETS_TO_CACHE = [
   './manifest.json',
   './favicon.svg',